chore: Rename vestigal heap/object.rs into object_entry.rs and remove unused functions
